Add Input.GetAxis as base for movement instead of explicit buttons Tweak acceleration and jumpVelocity values Restructure project into folders
