+++ Assets/Scripts/input.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 28d5549bf3e8d4574bc52cc4a9d63e54
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
