+++ Assets/Scripts/input/InputComponent.cs
@@ -5,9 +5,8 @@ using Unity.Entities;
 
 public struct InputComponent : IComponentData
 {
-    public int keyLeft;
-    public int keyRight;
-    public int keyUp;
+    public float horizontal;
+    public float vertical;
     public float mouseX;
     public float mouseY;
 }
