+++ Assets/Scripts/input/InputComponent.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: d536447532c3ad84c8b4fdd7d5521263
+guid: 9d9d2b433037048febc0b26919769dbe
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
