+++ Assets/Scripts/input/InputSystem.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: e0e249d4903d1144fafbb8a0772efa5b
+guid: e2254965e47634b30bce70cd1d1086b4
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
