+++ Assets/Scripts/movement/FrictionComponent.cs
