+++ Assets/Scripts/movement/FrictionComponent.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 5db30cafabc9e584b9516944e3e57922
+guid: a8287f04788224b77a7a9bc0693ec5d0
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
