+++ Assets/Scripts/movement/FrictionSystem.cs
