+++ Assets/Scripts/movement/FrictionSystem.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 74c1ee49a8475c04e85b22ae35b507a4
+guid: 7b836cc90f8e54aa88baa285f75807f9
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
