+++ Assets/Scripts/movement/MoveSpeedComponent.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 43f1daf6f468fae4f96f7a292f92e6a2
+guid: e3473da0c2ac8464d8f4bb38a56fbbd8
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
