+++ Assets/Scripts/movement/RevolveComponent.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 67b8a257a63ac6b46b34d72c35c9e66e
+guid: 6c6c340ff65cb4e3b97a334eb237be45
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
