+++ Assets/Scripts/physics/GravityComponent.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 2243edd5f79daec42af43a8bfd5bda5e
+guid: 6b827ac7d78eb469c9b01be825460b1f
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
